requirements.txt
fastapi
uvicorn[standard]
requests
loguru
pydantic
python-multipart
numpy
faiss-cpu
sentence-transformers
torch
python-dotenv
PyPDF2
python-docx
pandas
openpyxl
python-pptx
Pillow
pytesseract
beautifulsoup4